Fixed Vite template syntax issues
